Fixing STAT table and updating to UFR
The STAT table in this font was not set correctly—this has been corrected for proper rendering in Mac Word.

Additionally, the repository has been updated to follow the UFR guidelines, including building from a Makefile, and Continuous Integration, building directly from the designspace files and UFOs.

Checked with fontbakery and got cupcake!